updated the header for pubtheme modal
